Moved Dive Computer Rsc to component root
