+++ addons/main/Dialog.hpp
